Collection component, start of DB restructure
