+++ db/layout1.js
@@ -1,0 +1,22 @@
+const seedLayout1 = [
+  {
+    componentName: "NavBar",
+    props: {
+      companyName: "Dirty Dogs",
+      links: [
+        { href: "#Gallery", content: "Woof" },
+        { href: "#Contact", content: "Bark" },
+      ],
+    },
+  },
+  {
+    componentName: "Hero",
+    props: {
+      title: "Woof",
+      body: "woof woof woof woof woof woof woof bark bark woof woof woof woof woof woof woof woof bark bark woof woof woof woof woof woof woof woof bark bark woof",
+      imgSrc: "./assets/HappyPup.jpeg",
+    },
+  },
+];
+
+module.exports = seedLayout1;
