+++ db/seed.js
@@ -1,4 +1,6 @@
 const { client, createUser, createLayout } = require("./index");
+const seedLayout1 = require("./layout1");
+const seedLayout2 = require("./layout2");
 
 const seedUsers = [
   { username: "Jim", password: "password" },
@@ -7,65 +9,11 @@ const seedUsers = [
   { username: "Vincent", password: "gothgirls" },
 ];
 
-const seedLayout1 = [
-  {
-    componentName: "NavBar",
-    props: {
-      companyName: "Dirty Dogs",
-      links: [
-        { href: "#Gallery", content: "Woof" },
-        { href: "#Contact", content: "Bark" },
-      ],
-    },
-  },
-  {
-    componentName: "Hero",
-    props: {
-      title: "Woof",
-      body: "woof woof woof woof woof woof woof bark bark woof woof woof woof woof woof woof woof bark bark woof woof woof woof woof woof woof woof bark bark woof",
-      imgSrc: "./assets/HappyPup.jpeg",
-    },
-  },
-];
-
-const seedLayout2 = [
-  {
-    componentName: "NavBar",
-    props: {
-      companyName: "Catz",
-      links: [
-        { href: "#hero", content: "Meow" },
-        { href: "#about", content: "Hiss" },
-      ],
-    },
-  },
-  {
-    componentName: "Hero",
-    props: {
-      title: "MEOW",
-      body: "Meow meow meow meow meow meow meow meow meow meow meow",
-      imgSrc: "./assets/businessCat.jpeg",
-      buttons: [{ content: "Try Now!" }, { content: "Contact Us" }],
-    },
-  },
-  {
-    componentName: "Collection",
-    props: {
-      title: "Our Team",
-      desc: "The greatest team you've ever seen.",
-      collectionItems: [
-        { imgSrc: "./assets/businessCat.jpeg", itemTitle: "Bart", itemDesc: "CEO" },
-        { imgSrc: "./assets/businessCat.jpeg", itemTitle: "Tuna", itemDesc: "Accounting" },
-        { imgSrc: "./assets/businessCat.jpeg", itemTitle: "Chad", itemDesc: "IT" },
-      ],
-    },
-  },
-];
-
 const dropTables = async () => {
   try {
     console.log("Starting to drop tables!");
     await client.query(`
+        DROP TABLE IF EXISTS modules;
         DROP TABLE IF EXISTS siteLayouts;
         DROP TABLE IF EXISTS users;
     `);
@@ -75,18 +23,25 @@ const dropTables = async () => {
   }
 };
 
+//Leave siteLayouts for now until new structure is confirmed.
 const createTables = async () => {
   try {
     console.log("Creating tables!");
     await client.query(`
-        CREATE TABLE users (
-            id serial PRIMARY KEY,
-            username VARCHAR(255) UNIQUE NOT NULL,
-            password VARCHAR(255) NOT NULL
-            );
         CREATE TABLE siteLayouts (
           id serial PRIMARY KEY,
           layout jsonb NOT NULL
+        );
+        CREATE TABLE sites (
+          siteID serial PRIMARY KEY,
+          name VARCHAR(255) NOT NULL
+        );
+        CREATE TABLE modules (
+          moduleID serial PRIMARY KEY,
+          module jsonb NOT NULL,
+          sortOrder INT NOT NULL,
+          siteID INT REFERENCES sites(siteID) NOT NULL,
+          UNIQUE (sortOrder, siteID) 
         );
     `);
     console.log("Finished creating tables!");
@@ -99,7 +54,6 @@ const rebuildDB = async () => {
   client.connect();
   await dropTables();
   await createTables();
-  await Promise.all(seedUsers.map((user) => createUser(user.username, user.password)));
   await createLayout(seedLayout1);
   await createLayout(seedLayout2);
   client.end();
